+++ tests/test_simulation.py
@@ -11,23 +11,23 @@ from ruamel import yaml
 
 def test_compute_element_declaration() -> None:
     test_decl = ComputeElementDecl(
-            Identifier('test'), Reference('ns.model'))
+            Reference('test'), Reference('ns.model'))
     assert str(test_decl.name) == 'test'
     assert str(test_decl.implementation) == 'ns.model'
-    assert test_decl.count == 1
-    assert str(test_decl) == 'test[1]'
+    assert test_decl.multiplicity == []
+    assert str(test_decl) == 'test'
 
     test_decl = ComputeElementDecl(
-            Identifier('test'), Reference('ns2.model2'), 2)
-    assert isinstance(test_decl.name, Identifier)
+            Reference('test'), Reference('ns2.model2'), [1, 2])
+    assert isinstance(test_decl.name, Reference)
     assert str(test_decl.name) == 'test'
     assert str(test_decl.implementation) == 'ns2.model2'
-    assert test_decl.count == 2
-    assert str(test_decl) == 'test[2]'
+    assert test_decl.multiplicity == [1, 2]
+    assert str(test_decl) == 'test[1][2]'
 
     with pytest.raises(ValueError):
         test_decl = ComputeElementDecl(
-                Identifier('test'), Reference('ns2.model2[1]'))
+                Reference('test'), Reference('ns2.model2[1]'))
 
 
 def test_conduit() -> None:
@@ -53,12 +53,23 @@ def test_conduit() -> None:
     with pytest.raises(ValueError):
         Conduit(Reference('x[3].y.z'), test_ref)
 
+    test_conduit2 = Conduit(test_ref, test_ref2)
+    assert test_conduit == test_conduit2
+
+    test_conduit3 = Conduit(Reference('submodel1.port1'),
+                            Reference('submodel2.port2'))
+    assert test_conduit == test_conduit3
+
+    assert 'Conduit' in str(test_conduit)
+    assert 'submodel1.port1' in str(test_conduit)
+    assert 'submodel2.port2' in str(test_conduit)
+
 
 def test_simulation() -> None:
     macro = ComputeElementDecl(
-            Identifier('macro'), Reference('my.macro'))
+            Reference('macro'), Reference('my.macro'))
     micro = ComputeElementDecl(
-            Identifier('micro'), Reference('my.micro'))
+            Reference('micro'), Reference('my.micro'))
     comp_els = [macro, micro]
     conduit1 = Conduit(Reference('macro.intermediate_state'),
                        Reference('micro.initial_state'))
@@ -116,8 +127,8 @@ def test_dump_simulation() -> None:
 
     impl1 = Reference('test.impl1')
     impl2 = Reference('test.impl2')
-    ce1 = ComputeElementDecl(Identifier('ce1'), impl1)
-    ce2 = ComputeElementDecl(Identifier('ce2'), impl2)
+    ce1 = ComputeElementDecl(Reference('ce1'), impl1)
+    ce2 = ComputeElementDecl(Reference('ce2'), impl2)
     ce1_out = Reference('ce1.state_out')
     ce2_in = Reference('ce2.init_in')
     ce2_out = Reference('ce2.fini_out')
@@ -130,12 +141,8 @@ def test_dump_simulation() -> None:
     print(text)
     assert text == ('name: test_sim\n'
                     'compute_elements:\n'
-                    '  ce1:\n'
-                    '    implementation: test.impl1\n'
-                    '    count: 1\n'
-                    '  ce2:\n'
-                    '    implementation: test.impl2\n'
-                    '    count: 1\n'
+                    '  ce1: test.impl1\n'
+                    '  ce2: test.impl2\n'
                     'conduits:\n'
                     '  ce1.state_out: ce2.init_in\n'
                     '  ce2.fini_out: ce1.boundary_in\n'