+++ CHANGELOG.rst
@@ -5,6 +5,21 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.5.0
+*****
+
+Changed
+-------
+
+* Add multiplicity to ComputeElementDecl
+
+Fixed
+-----
+
+* ComputeElementDecl has a Reference for its name
+* Savorizing issue in ComputeElementDecl
+
+
 0.4.0
 *****